Create Permutations
This is a program about all permutation of elements in the string. For solving this problem I've used the tree of permutation, so firstly we permute first elements, then for each option we use our function again, this is done until it comes down to the permutation of the last element, so this is recursion.